+++ src/Intermediate/Exercise1/Suma.java
@@ -1,18 +1,18 @@
-package Intermediate.Ejercicio1;
+package Intermediate.Exercise1;
 
 /**
  *
  * @author dvchinx
  */
-public class Multp {
+public class Suma {
 
     protected float a, b;
 
-    public Multp() {
+    public Suma() {
     }
 
     public int Operacion(float a, float b) {
-        return (int) Math.round(a * b);
+        return (int) Math.round(a + b);
     }
 
     public float getA() {
